pyproject.toml
[project]
name = "fitbit-to-kml"
version = "0.1.0"
description = "A collection of tools to dump FitBit activity data and convert it to KML format for further analysis and visualization."
readme = "readme.md"
requires-python = ">=3.14"
dependencies = [
    "pytest>=9.0.0",
    "python-tcxparser>=2.4.0",
    "requests>=2.32.0",
    "requests-oauth2client>=1.7.0",
    "simplekml>=1.3.6",
    "structlog>=25.5.0",
]

[dependency-groups]
dev = ["pytest>=9.0.0"]
